+++ packages/desktop/src/components/settings/SettingsPanel.spec.ts
@@ -4,7 +4,7 @@ import { render } from "vitest-browser-svelte";
 import { mockIPC } from "@tauri-apps/api/mocks";
 import { accountStore } from "$lib/stores/account.svelte";
 import { categoryStore } from "$lib/stores/categories.svelte";
-import type { Category } from "$lib/types";
+import type { Category, Currency } from "$lib/types";
 
 beforeEach(() => {
     categoryStore.clear();
@@ -15,7 +15,25 @@ mockIPC((cmd) => {
         return { currencyCode: "USD" };
     }
     if (cmd === "currencies") {
-        return ["USD", "CAD", "ZAR", "ZMW", "TSH"];
+        const currencies: Currency[] = [
+            {
+                code:"AUD",
+                name: ""
+            },
+            {
+                code:"ZMW",
+                name: ""
+            },
+            {
+                code:"CAD",
+                name: ""
+            },
+            {
+                code:"USD",
+                name: ""
+            }
+        ];
+        return currencies;
     }
 });