+++ packages/desktop/src/components/sidebar/Sidebar.spec.ts
@@ -4,6 +4,7 @@ import { accountStore, mockCreateAccount } from "$lib/stores/account.svelte";
 import Sidebar from "./Sidebar.svelte";
 import { mockIPC } from "@tauri-apps/api/mocks";
 import { mockSettings, settingsStore } from "$lib/stores/settings.svelte";
+import type { Currency } from "$lib/types";
 
 mockIPC((cmd) => {
     if (cmd === "settings") {
@@ -13,17 +14,34 @@ mockIPC((cmd) => {
         return;
     }
     if (cmd === "currencies") {
-        return ["USD", "CAD", "ZAR", "ZMW", "TSH"];
+        const currencies: Currency[] = [
+            {
+                code:"AUD",
+                name: ""
+            },
+            {
+                code:"ZMW",
+                name: ""
+            },
+            {
+                code:"CAD",
+                name: ""
+            },
+            {
+                code:"USD",
+                name: ""
+            }
+        ];
+        return currencies;
     }
 });
 
 beforeEach(() => {
     accountStore.clear();
     settingsStore.reset();
-    // clearMocks();
 });
 
-describe("Navigation Panel",() => {
+describe("Sidebar",() => {
     test("has navigation links",async() => {
         const screen = await render(Sidebar);
         const transactionsLink = screen.getByRole("link",{ name:"Transactions" });
@@ -75,7 +93,6 @@ describe("Navigation Panel",() => {
         await expect.element(spendingLink).not.toBeInTheDocument();
     });
 
-    // TODO: maybe test reactivity
     test("account list", async () => {
         mockCreateAccount();
         await accountStore.createAccount({ name: "Account 1",startingBalance: "20.00" });
@@ -83,7 +100,7 @@ describe("Navigation Panel",() => {
         const screen = await render(Sidebar);
 
         await expect.element(screen.getByText("Account 1")).toBeVisible();
-        await expect.element(screen.getByText("$20.00")).toBeVisible();
+        await expect.element(screen.getByText("K20.00")).toBeVisible();
         expect(screen.getByRole("listitem").all()).toHaveLength(2);
     });
 
@@ -94,6 +111,6 @@ describe("Navigation Panel",() => {
         const screen = await render(Sidebar);
 
         await expect.element(screen.getByText("Accounts")).toBeVisible();
-        await expect.element(screen.getByText("$520.00")).toBeVisible();
+        await expect.element(screen.getByText("K520.00")).toBeVisible();
     });
 });